+++ CHANGELOG.md
@@ -40,6 +40,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Replace slice in templates with sprig substr. #1093
 - Fix an invalid format issue for the GitHub nightly build action. #1258
 - Return non-zero exit code on overlay build failure #1393
+- Return non-zero exit code on container copy failure #1377
 
 ## v4.5.8, unreleased